# opt: use constraint primitives for idxconstraint