--- a/pkg/sql/opt/idxconstraint/testdata/tuple
+++ b/pkg/sql/opt/idxconstraint/testdata/tuple
@@ -42,14 +42,12 @@ index-constraints vars=(int, int) index=(@1, @2)
 ----
 [/1/1 - /1/3]
 [/2/1 - /2/3]
-Remaining filter: @2 IN (1, 2, 3)
 
 index-constraints vars=(int, int) index=(@1 desc, @2 desc)
 @1 IN (1, 2) AND @2 IN (1, 2, 3)
 ----
 [/2/3 - /2/1]
 [/1/3 - /1/1]
-Remaining filter: @2 IN (1, 2, 3)
 
 index-constraints vars=(int, int) index=(@1, @2)
 @1 >= 2 AND @1 <= 4 AND @2 IN (1, 2, 3)