… pkg/sql/opt/constraint/constraint_test.go
… pkg/sql/opt/constraint/span_test.go
… pkg/sql/opt/constraint/spans_test.go
… pkg/sql/opt/constraint/utils_test.go
… pkg/sql/opt/idxconstraint/index_constraints.go
… pkg/sql/opt/idxconstraint/index_constraints_test.go
… pkg/sql/opt/idxconstraint/spans.go
… pkg/sql/opt/idxconstraint/spans_test.go
… pkg/sql/opt_index_selection_test.go