#### opt/idxconstraint: benchmarks with single long span

Adding two benchmarks that show the scaling when we have many columns,
but just one span.

#### opt: use constraint primitives for idxconstraint

This change reworks the index constraints code to use the new
primitives in `opt/constraint`.

Notable changes:
 - IndexColumnInfo now uses OrderingColumn instead of separate index
   and direction.
 - more efficient Columns.Equals implementation.
 - more efficient initializers for Span, Spans and Constraint. In
   particular, Span.Init no longer checks that the keys are in the
   expected order (this had a not-insignificant cost).
 - during filter simplification, instead of using CutFirstColumn and
   SubsetOf, we now generate the cut spans one at a time and use
   ContainsSpan. This is more efficient (we don't need to copy spans)
   and the necessary primitive is simpler.

Benchmark results:
```
name                                      old time/op    new time/op    delta
IndexConstraints/point-lookup-8              361ns ± 0%     288ns ± 2%   -20.25%  (p=0.000 n=8+10)
IndexConstraints/no-constraints-8            146ns ± 0%     171ns ± 4%   +17.06%  (p=0.000 n=9+10)
IndexConstraints/range-8                    1.67µs ± 1%    1.57µs ± 1%    -5.86%  (p=0.000 n=9+10)
IndexConstraints/range-2d-8                 2.86µs ± 3%    3.04µs ± 1%    +6.48%  (p=0.000 n=10+9)
IndexConstraints/many-columns-8             5.13µs ± 2%    6.05µs ± 2%   +18.04%  (p=0.000 n=9+10)
IndexConstraints/single-jumbo-span-10-8     16.5µs ± 2%    20.1µs ± 0%   +21.24%  (p=0.000 n=10+8)
IndexConstraints/single-jumbo-span-100-8     824µs ± 2%    1102µs ± 1%   +33.79%  (p=0.000 n=9+10)

name                                      old alloc/op   new alloc/op   delta
IndexConstraints/point-lookup-8               224B ± 0%       52B ± 0%   -76.79%  (p=0.000 n=10+10)
IndexConstraints/no-constraints-8             128B ± 0%       52B ± 0%   -59.38%  (p=0.000 n=10+10)
IndexConstraints/range-8                      584B ± 0%       64B ± 0%   -89.04%  (p=0.000 n=10+10)
IndexConstraints/range-2d-8                   912B ± 0%      136B ± 0%   -85.09%  (p=0.000 n=10+10)
IndexConstraints/many-columns-8             1.75kB ± 0%    0.89kB ± 0%   -49.32%  (p=0.000 n=10+10)
IndexConstraints/single-jumbo-span-10-8     4.88kB ± 0%    1.95kB ± 0%   -60.00%  (p=0.000 n=10+10)
IndexConstraints/single-jumbo-span-100-8     201kB ± 0%     171kB ± 0%   -14.79%  (p=0.000 n=10+10)

name                                      old allocs/op  new allocs/op  delta
IndexConstraints/point-lookup-8               4.00 ± 0%      2.00 ± 0%   -50.00%  (p=0.000 n=10+10)
IndexConstraints/no-constraints-8             1.00 ± 0%      2.00 ± 0%  +100.00%  (p=0.000 n=10+10)
IndexConstraints/range-8                      16.0 ± 0%       3.0 ± 0%   -81.25%  (p=0.000 n=10+10)
IndexConstraints/range-2d-8                   24.0 ± 0%       5.0 ± 0%   -79.17%  (p=0.000 n=10+10)
IndexConstraints/many-columns-8               36.0 ± 0%      15.0 ± 0%   -58.33%  (p=0.000 n=10+10)
IndexConstraints/single-jumbo-span-10-8       91.0 ± 0%      21.0 ± 0%   -76.92%  (p=0.000 n=10+10)
IndexConstraints/single-jumbo-span-100-8       901 ± 0%       201 ± 0%   -77.69%  (p=0.000 n=10+10)
```
Release note: None

Many thanks to @andy-kimball for tracking down various performance problems.