@andy-kimball 2018-03-26
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Review status: 0 of 20 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---

*[pkg/sql/opt/idxconstraint/index_constraints.go, line 255 at r2](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192#-L8XdJ0z9biuS1zhQQl-:-L8XdJ0z9biuS1zhQQl0:b-o0dvx4) ([raw file](https://github.com/cockroachdb/cockroach/blob/0b9c16a95bb399023b8024454764613ddeafe1eb/pkg/sql/opt/idxconstraint/index_constraints.go#L255)):*
> ```Go
> 
> 	case opt.NeOp, opt.IsNotOp:
> 		startKey, startBoundary := c.notNullStartKey(offset)
> ```

A comment here would be good, maybe something like:
```
// Build constraint that doesn't contain the key:
//   if nullable    : [ - key) (key - ]
//   if not nullable: (NULL - key) (key - ]
```

---


*Comments from [Reviewable](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192#-:-L8XfVJN4-zyc19-f1DH:bnfp4nl)*
<!-- Sent from Reviewable.io -->