@RaduBerinde 2018-03-26
Review status: 0 of 20 files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[pkg/sql/opt/idxconstraint/index_constraints.go, line 255 at r2](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192#-L8XdJ0z9biuS1zhQQl-:-L8Xh6kaBIBKAoGh1EBf:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/0b9c16a95bb399023b8024454764613ddeafe1eb/pkg/sql/opt/idxconstraint/index_constraints.go#L255)):*
<details><summary><i>Previously, andy-kimball (Andy Kimball) wrote…</i></summary><blockquote>

A comment here would be good, maybe something like:
```
// Build constraint that doesn't contain the key:
//   if nullable    : [ - key) (key - ]
//   if not nullable: (NULL - key) (key - ]
```
</blockquote></details>

Done.

---


*Comments from [Reviewable](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192)*
<!-- Sent from Reviewable.io -->