@rytaft 2018-03-26
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Reviewed 1 of 1 files at r1, 20 of 20 files at r2, 1 of 1 files at r3.
Review status: all files reviewed at latest revision, 1 unresolved discussion, some commit checks failed.

---

*[pkg/sql/opt_index_selection.go, line 223 at r2](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192#-L8Xc5xI1q5kx9tidYL7:-L8Xc5xI1q5kx9tidYL8:b-49rg34) ([raw file](https://github.com/cockroachdb/cockroach/blob/0b9c16a95bb399023b8024454764613ddeafe1eb/pkg/sql/opt_index_selection.go#L223)):*
> ```Go
> 		for i, c := range candidates {
> 			log.Infof(
> 				ctx, "%d: selectIndex(%s): cost=%v spans=%s reverse=%t",
> ```

should this be `constraint=%s`?

---

*[pkg/sql/opt_index_selection.go, line 596 at r2](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192#-L8XdPHI-vyQTnUxjhr5:-L8XdPHI-vyQTnUxjhr6:b6aan7j) ([raw file](https://github.com/cockroachdb/cockroach/blob/0b9c16a95bb399023b8024454764613ddeafe1eb/pkg/sql/opt_index_selection.go#L596)):*
> ```Go
> }
> 
> // spanFromConstraintSpan converts an constraint.Span to a
> ```

[nit] an -> a

---

*[pkg/sql/opt/constraint/span.go, line 86 at r2](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192#-L8Xi62q-3dKq3IJVe9G:-L8Xi62q-3dKq3IJVe9H:b-64qcr5) ([raw file](https://github.com/cockroachdb/cockroach/blob/0b9c16a95bb399023b8024454764613ddeafe1eb/pkg/sql/opt/constraint/span.go#L86)):*
> ```Go
> }
> 
> // EndBoundary returns whether the start key is included or excluded.
> ```

start -> end

---

*[pkg/sql/opt/constraint/spans.go, line 36 at r2](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192#-L8Xir3z-f6eGIAcMbTm:-L8Xir4-1hCYkzDJwjAb:b2x08ic) ([raw file](https://github.com/cockroachdb/cockroach/blob/0b9c16a95bb399023b8024454764613ddeafe1eb/pkg/sql/opt/constraint/spans.go#L36)):*
> ```Go
> 
> // Alloc allocates enough space to support the given amount of spans without
> // reallocation. Does nothing if the structure already contains spans.
> ```

Doesn't look like that's true - it checks if `s.numSpans > 0`

---

*[pkg/sql/opt/idxconstraint/index_constraints.go, line 437 at r2](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192#-L8XnVt1Bmw_8R7Bdlji:-L8XnVt1Bmw_8R7Bdljj:baho6z1) ([raw file](https://github.com/cockroachdb/cockroach/blob/0b9c16a95bb399023b8024454764613ddeafe1eb/pkg/sql/opt/idxconstraint/index_constraints.go#L437)):*
> ```Go
> 	}
> 
> 	// We use makeNotNullSpan to disallow NULLs on the first column.
> ```

update comment

---


*Comments from [Reviewable](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192#-:-L8Xrcms1pZv59NRP5_p:bnfp4nl)*
<!-- Sent from Reviewable.io -->