@RaduBerinde 2018-03-26
TFTRs!

---

Review status: all files reviewed at latest revision, 6 unresolved discussions, some commit checks failed.

---

*[pkg/sql/opt_index_selection.go, line 223 at r2](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192#-L8Xc5xI1q5kx9tidYL7:-L8Xz89AAI9VggaY20Da:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/0b9c16a95bb399023b8024454764613ddeafe1eb/pkg/sql/opt_index_selection.go#L223)):*
<details><summary><i>Previously, rytaft wrote…</i></summary><blockquote>

should this be `constraint=%s`?
</blockquote></details>

Done.

---

*[pkg/sql/opt_index_selection.go, line 596 at r2](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192#-L8XdPHI-vyQTnUxjhr5:-L8XzAMo7zonejrQLTqY:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/0b9c16a95bb399023b8024454764613ddeafe1eb/pkg/sql/opt_index_selection.go#L596)):*
<details><summary><i>Previously, rytaft wrote…</i></summary><blockquote>

[nit] an -> a
</blockquote></details>

Done.

---

*[pkg/sql/opt/constraint/span.go, line 86 at r2](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192#-L8Xi62q-3dKq3IJVe9G:-L8XzEJ8BqhLtH-ifepw:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/0b9c16a95bb399023b8024454764613ddeafe1eb/pkg/sql/opt/constraint/span.go#L86)):*
<details><summary><i>Previously, rytaft wrote…</i></summary><blockquote>

start -> end
</blockquote></details>

Done.

---

*[pkg/sql/opt/constraint/spans.go, line 36 at r2](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192#-L8Xir3z-f6eGIAcMbTm:-L8Xyt9g9ymEhRr1USkO:bjf7f3k) ([raw file](https://github.com/cockroachdb/cockroach/blob/0b9c16a95bb399023b8024454764613ddeafe1eb/pkg/sql/opt/constraint/spans.go#L36)):*
<details><summary><i>Previously, rytaft wrote…</i></summary><blockquote>

Doesn't look like that's true - it checks if `s.numSpans > 0`
</blockquote></details>

Good catch!  Fixed, and added a test.

---

*[pkg/sql/opt/idxconstraint/index_constraints.go, line 437 at r2](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192#-L8XnVt1Bmw_8R7Bdlji:-L8XzM-8DMYcKIFrSj5z:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/0b9c16a95bb399023b8024454764613ddeafe1eb/pkg/sql/opt/idxconstraint/index_constraints.go#L437)):*
<details><summary><i>Previously, rytaft wrote…</i></summary><blockquote>

update comment
</blockquote></details>

Done.

---


*Comments from [Reviewable](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192#-:-L8XzNco8hF34AHdTXQt:bz5mair)*
<!-- Sent from Reviewable.io -->