@rytaft 2018-03-26
Reviewed 5 of 5 files at r4.
Review status: all files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---



*Comments from [Reviewable](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24192)*
<!-- Sent from Reviewable.io -->